BREAKING: introduce template blocks to config.hcl
The "default_template" field has been removed from "source" blocks and
has been replaced with a "template" block. To specify a default
template, create a template with a name of "default" like so:

    source "command" "gh" {
      command = "gh repo list --json nameWithOwner"

      template {
        name = "default"
        value = "{{.nameWithOwner}}"
      }
    }

Templates with many different names can be specified, this is will be
integrated into the CLI in a following commit.